# packages/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Hinter Factory ML API
  version: 0.1.0
  description: >-
    Every scoped endpoint (documents, tags, labeling functions, LF runs,
    gold labels, probabilistic labels, evaluation) requires a `project_id`
    query parameter (or, for multipart uploads, a `project_id` form field).
    There is no implicit fallback project; calls that omit it return 400.
    Use `GET /v1/projects` to discover available projects and
    `POST /v1/projects` to create one. The web UI threads `project_id`
    through every request automatically via its project context.
servers:
  - url: http://localhost:8000
...
paths:
...
  /v1/documents/upload:
    post:
      tags: [documents]
      operationId: uploadDocuments
      summary: Ingest documents from CSV or JSON
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                text_column:
                  type: string
                  default: text
                  description: CSV column name for document body
                id_column:
                  type: string
                  description: Optional CSV column to use as stable id
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentIngestResult"
  /v1/documents:
    get:
      tags: [documents]
      operationId: listDocuments
      summary: Explore corpus with search and facets
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: length_bucket
          style: form
          explode: true
          schema:
            type: array
            items:
              $ref: "#/components/schemas/LengthBucket"
        - in: query
          name: metadata_key
          schema: { type: string }
        - in: query
          name: metadata_value
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
  /v1/documents/{document_id}:
    get:
      tags: [documents]
      operationId: getDocument
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
...
  /v1/documents/facets/metadata-values:
    get:
      tags: [documents]
      operationId: listMetadataFacetValues
      parameters:
        - in: query
          name: key
          required: true
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { type: string }
...
  /v1/labeling-functions:
    get:
      tags: [labelingFunctions]
      operationId: listLabelingFunctions
      parameters:
        - in: query
          name: tag_id
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LabelingFunction"
    post:
      tags: [labelingFunctions]
      operationId: createLabelingFunction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabelingFunctionCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabelingFunction"
  /v1/labeling-functions/{labeling_function_id}:
    patch:
      tags: [labelingFunctions]
      operationId: updateLabelingFunction
      parameters:
        - $ref: "#/components/parameters/LabelingFunctionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabelingFunctionUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabelingFunction"
    delete:
      tags: [labelingFunctions]
      operationId: deleteLabelingFunction
      parameters:
        - $ref: "#/components/parameters/LabelingFunctionId"
      responses:
        "204":
          description: Deleted
  /v1/labeling-functions/{labeling_function_id}/preview:
    post:
      tags: [labelingFunctions]
      operationId: previewLabelingFunction
      parameters:
        - $ref: "#/components/parameters/LabelingFunctionId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabelingFunctionPreviewRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabelingFunctionPreviewResponse"
...
  /v1/gold-labels:
    get:
      tags: [goldLabels]
      operationId: listGoldLabels
      parameters:
        - in: query
          name: document_id
          schema: { type: string }
        - in: query
          name: document_ids
          description: Repeat query key for each id (e.g. document_ids=a&document_ids=b)
          schema:
            type: array
            items: { type: string }
          style: form
          explode: true
        - in: query
          name: tag_id
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GoldLabel"
    post:
      tags: [goldLabels]
      operationId: createGoldLabel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoldLabelCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoldLabel"
components:
  parameters:
    DocumentId:
      name: document_id
      in: path
      required: true
      schema: { type: string }
    LabelingFunctionId:
      name: labeling_function_id
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    LengthBucket:
      type: string
      enum: [short, medium, long]
      description: short <100 chars, medium 100-499, long 500+
    Document:
      type: object
      required: [id, text, metadata, char_length, created_at]
      properties:
        id: { type: string }
        project_id: { type: string }
        text: { type: string }
        metadata:
          type: object
          additionalProperties: true
        char_length: { type: integer }
        created_at: { type: string, format: date-time }
    DocumentListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Document"
        total: { type: integer }
    DocumentIngestResult:
      type: object
      required: [inserted, skipped, errors]
      properties:
        inserted: { type: integer }
        skipped: { type: integer }
        errors:
          type: array
          items: { type: string }
...
    LabelingFunctionType:
      type: string
      enum: [regex, keywords, structural, zeroshot, llm_prompt]
    LabelingFunction:
      type: object
      required: [id, tag_id, name, type, config, enabled, created_at]
      properties:
        id: { type: string }
        tag_id: { type: string }
        name: { type: string }
        type:
          $ref: "#/components/schemas/LabelingFunctionType"
        config:
          type: object
          additionalProperties: true
          description: Type-specific JSON (patterns, thresholds, etc.)
        enabled: { type: boolean }
        created_at: { type: string, format: date-time }
    LabelingFunctionCreate:
      type: object
      required: [tag_id, name, type, config]
      properties:
        tag_id: { type: string }
        name: { type: string }
        type:
          $ref: "#/components/schemas/LabelingFunctionType"
        config:
          type: object
          additionalProperties: true
        enabled: { type: boolean, default: true }
    LabelingFunctionUpdate:
      type: object
      properties:
        name: { type: string }
        config:
          type: object
          additionalProperties: true
        enabled: { type: boolean }
    LabelingFunctionPreviewRequest:
      type: object
      properties:
        limit: { type: integer, minimum: 1, maximum: 200, default: 25 }
        document_ids:
          type: array
          items: { type: string }
    LabelingFunctionPreviewRow:
      type: object
      required: [document_id, vote, text_preview]
      properties:
        document_id: { type: string }
        vote: { $ref: "#/components/schemas/LfVote" }
        text_preview: { type: string }
    LabelingFunctionPreviewResponse:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/LabelingFunctionPreviewRow"
    LfVote:
      type: integer
      enum: [-1, 0, 1]
      description: 1 positive for tag, -1 negative, 0 abstain
...
    GoldLabel:
      type: object
      required: [id, document_id, tag_id, value, created_at]
      properties:
        id: { type: string }
        document_id: { type: string }
        tag_id: { type: string }
        value:
          type: integer
          enum: [-1, 0, 1]
          description: Gold / manual label aligned with LF votes (1 positive, -1 negative, 0 abstain)
        note: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    GoldLabelCreate:
      type: object
      required: [document_id, tag_id, value]
      properties:
        document_id: { type: string }
        tag_id: { type: string }
        value:
          type: integer
          enum: [-1, 0, 1]
        note: { type: string, nullable: true }
...